Properly grouped code
