+++ bin/micro.js
@@ -6,6 +6,10 @@ const path = require('path')
 // Packages
 const parseArgs = require('nanomist')
 
+// Utilities
+const serve = require('../lib')
+const getModule = require('../lib/module')
+
 // Check if the user defined any options
 const flags = parseArgs(process.argv, {
   string: ['host', 'port'],
@@ -41,4 +45,32 @@ if (file[0] !== '/') {
   file = path.resolve(process.cwd(), file)
 }
 
-require('../lib')(file, flags)
+const loadedModule = getModule(file)
+const server = serve(loadedModule)
+
+let host = flags.host
+
+if (host === '0.0.0.0') {
+  host = null
+}
+
+server.on('error', err => {
+  console.error('micro:', err.stack)
+
+  // eslint-disable-next-line unicorn/no-process-exit
+  process.exit(1)
+})
+
+server.listen(flags.port, host, () => {
+  const details = server.address()
+  const url = `http://localhost:${details.port}`
+
+  process.on('SIGINT', () => {
+    server.close()
+    process.exit(0)
+  })
+
+  if (!process.env.NOW) {
+    console.log(`Micro is running: ${url}`)
+  }
+})
