+++ package.json
@@ -2,7 +2,7 @@
   "name": "micro",
   "version": "7.3.3",
   "description": "Asynchronous HTTP microservices",
-  "main": "./lib/server.js",
+  "main": "./lib/index.js",
   "files": [
     "bin",
     "lib"
