+++ test/development.js
@@ -4,7 +4,7 @@ const request = require('request-promise')
 const listen = require('test-listen')
 
 process.env.NODE_ENV = 'development'
-const micro = require('../lib/server')
+const micro = require('../')
 
 const getUrl = fn => {
   const srv = micro(fn)
