+++ test/index.js
@@ -4,7 +4,7 @@ const request = require('request-promise')
 const sleep = require('then-sleep')
 const resumer = require('resumer')
 const listen = require('test-listen')
-const micro = require('../lib/server')
+const micro = require('../')
 
 const { send, sendError, buffer, json } = micro
